FEXCore: Move `JITSymbolBuffer` to internal header
This definition doesn't need to be exposed in the public API.
